Add `MockMountpointS3Client` and associated unit tests
`MockMountpointS3Client` has a similar interface to
`MountpointS3Client`, except it has `add_object` and
`remove_object` functions which bypass the process of
PUT/DELETE. This makes it easier for us to test S3 methods in
isolation.
Python and Rust tests are changed to be unit tests rather than integration tests.