Don't try to deallocate clients after a fork
The destructor path for the CRT tries to join on threads that aren't
available, which causes all kinds of fun, including segfaults. Strictly
speaking this is now a memory leak, but it's only going to happen when
clients are torn down, which we expect should be fairly rare as PyTorch
goes to some lengths to have its loader processes live a long time.

Signed-off-by: James Bornholt <bornholt@amazon.com>