Avoid excessive memory consumption when utilizing s3map_dataset  (#87)
The goal of the changes is to minimize the retention time for S3 data loaded in memory. 
This was achieved by delaying the creation of S3Readers until right before reading the given object. 
Immediately after reading finishes, ownership of the S3Reader containing the loaded content is 
transferred to the calling client.
* Don't create instance of S3Reader until it needed and transfer its ownership to decrease memory pressure.
* Simplify S3BucketKey with NamedTuple
* Format files according to the Black lint
* Make list_objects_from_prefix return iterable instead of list
* Make get_objects_from_uris return iterable instead of list
* Restore test_get_objects_from_uris_fail previous behavior
* Parse URIs eagerly, then return everything lazily.
* Rename list_objects_from_prefix to get_objects_from_prefix

---------

Co-authored-by: Ilya Isaev <iisaev@amazon.co.uk>