fix(test): remove unnecessary test in test_logging_to_file
In test_logging_to_file debug,awscrt=off case, we were previously testing for existence of awscrt=off in STDOUT, but in reality that was simply part of the test name that became bucket name through request.node.name. Now that we are reusing the image_directory fixture, that is no longer a suitable test so I removed it.